+++ package.json
@@ -55,6 +55,7 @@
     "tailwindcss": "^3.2.4",
     "typescript": "~4.7.4",
     "vite": "^3.2.0",
+    "vite-plugin-vue-type-imports": "^0.2.4",
     "vite-svg-loader": "^4.0.0",
     "vitest": "^0.25.6",
     "vue-eslint-parser": "^9.1.0",
