+++ src/components/baseButton/uses.ts
@@ -1,9 +1,9 @@
 import { computed } from 'vue';
 
-export type ButtonNativeType = 'button' | 'reset' | 'submit';
-export type ButtonType = 'primary' | 'secondary' | 'success' | 'info' | 'warning' | 'error';
+type ButtonNativeType = 'button' | 'reset' | 'submit';
+type ButtonType = 'primary' | 'secondary' | 'success' | 'info' | 'warning' | 'error';
 
-export interface Props {
+interface Props {
   nativeType: ButtonNativeType;
   type: ButtonType;
 }
@@ -28,4 +28,10 @@ const useButton = (props: Props, emit: Emit) => {
 
 export {
   useButton,
+};
+export type {
+  ButtonNativeType,
+  ButtonType,
+  Props,
+  Emit,
 };
