+++ src/components/buttonIcon/ButtonIcon.vue
@@ -10,17 +10,9 @@
 </template>
 
 <script setup lang="ts">
-import type { ButtonNativeType, ButtonType } from '@/components/baseButton/uses';
 import BaseButton from '@/components/baseButton/BaseButton.vue';
 import { useButtonIcon } from './uses';
-
-interface Props {
-  nativeType?: ButtonNativeType;
-  type?: ButtonType;
-}
-interface Emit {
-  (e: 'click'): void;
-}
+import type { Props, Emit } from './uses';
 
 const props = withDefaults(defineProps<Props>(), {
   nativeType: 'button',
