+++ src/components/cardText/CardTextSkeleton.vue
@@ -7,7 +7,6 @@ interface Props {
 }
 
 defineProps<Props>();
-
 </script>
 
 <template>
